refactor(db): rename oauthSessions to OAuthSessions.ts
