+++ packages/db/src/index.ts
@@ -2,7 +2,7 @@ export * from "./apiKeys.js";
 export * from "./adminAuth.js";
 export * from "./db.js";
 export * from "./logs.js";
-export * from "./oauthSessions.js";
+export * from "./OAuthSessions.js";
 export * from "./providers.js";
 export * from "./quota.js";
 export * from "./settings.js";
